Matriz dispersa
Terminado, solo falta los 2 reportes a consola, y resolver posibles errores que puedan pasar en el programa

+++ Proyecto-Fase 1/Estructuras/Cola/Servicios.cs
@@ -7,10 +7,10 @@ namespace List
         public int id_Repuesto { get; set; }
         public int id_Vehiculo { get; set; }
         public string detalles { get; set; }
-        public int costo { get; set; }
+        public double costo { get; set; }
 
         //Constructor
-        public Servicios(int ID, int id_Repair, int id_Vehicle, string details, int cost)
+        public Servicios(int ID, int id_Repair, int id_Vehicle, string details, double cost)
         {
             id = ID;
             id_Repuesto = id_Repair;
